Add: Custom hooks for main-page (useCities, useOffers)
