+++ src/components/app/app.tsx
@@ -4,16 +4,16 @@ import FavoritesPage from '../../pages/favorites-page/favorites-page';
 import OfferPage from '../../pages/offer-page/offer-page';
 import LoginPage from '../../pages/login-page/login-page';
 import Page404 from '../../pages/page-404/page-404';
-import { PrivateRoute, PublicRoute } from '../../pages/access-route/access-route';
 import { AppRoute } from '../../constants';
+import { PrivateRoute, PublicRoute } from '../../pages/access-route/access-route';
 import { Route, Routes } from 'react-router-dom';
-
-import './stable-width.css';
 import { useAppSelector } from '../../hooks';
 import browserHistory from '../../browser-history';
 import HistoryRouter from '../history-route/history-route';
 import { getAuthorizationStatus } from '../../store/user/user.selectors';
 import ScrollToTop from '../scroll-to-top/scroll-to-top';
+
+import './stable-width.css';
 
 function App(): React.JSX.Element {
 	const authorizationStatus = useAppSelector(getAuthorizationStatus);
