+++ src/pages/main-page/main-page.tsx
@@ -3,56 +3,30 @@ import LocationsList from '../../components/location-list/location-list';
 import OfferList from '../../components/offer-list/offer-list';
 import Sort from '../../components/sort/sort';
 import LeafletMap from '../../components/leaflet-map/leaflet-map';
-import { useActionCreators, useAppSelector, useDocumentTitle } from '../../hooks';
-import { CITIES } from '../../constants';
+import { useAppSelector, useCities, useDocumentTitle } from '../../hooks';
 import clsx from 'clsx';
-import { DEFAULT_SORT, RequestStatus, SortMethod } from '../../constants/common';
+import { RequestStatus } from '../../constants/common';
 import LoadingScreen from '../loading-screen/loading-screen';
 import ErrorElement, { ErrorMessage } from '../../components/error-element/error-element';
 import { ErrorCause } from '../../constants/errors';
 import { toast } from 'react-toastify';
-import { getAllOffers, getAllOffersFetchingStatus, getCity, getSort } from '../../store/offers/offers.selectors';
-import { SortMap, getPluralPlaces } from '../../utils/convert';
-import { useSearchParams } from 'react-router-dom';
-import { offersActions } from '../../store/offers/offers.slice';
-import { useEffect } from 'react';
+import { getAllOffersFetchingStatus } from '../../store/offers/offers.selectors';
+import { getPluralPlaces } from '../../utils/convert';
+import { useOffers } from '../../hooks/use-offers/use-offers';
 
 /**
  * Компонент главного экрана
  */
 function MainPage(): React.JSX.Element {
 	useDocumentTitle('Main');
-	const {setCity, setSort} = useActionCreators(offersActions);
-
-	const cities = Array.from(CITIES);
-	const currentCity = useAppSelector(getCity);
-	const currenSort = useAppSelector(getSort);
-
+	const {cities, currentCity, currentSort } = useCities();
+	const {offers, isEmpty} = useOffers(currentCity, currentSort);
 	const offersLoadedStatus = useAppSelector(getAllOffersFetchingStatus);
-	const rawOffers = useAppSelector(getAllOffers).slice();
-	const offers = rawOffers
-		.filter((offer) => offer.city.name === currentCity)
-		.sort(SortMap[currenSort].sortFunc);
-	const isEmpty = offers.length === 0;
-
-	const [searchParams] = useSearchParams();
-	const initialCity = searchParams.get('filter') || cities[0];
-	const initialSort = searchParams.get('sort') || SortMethod[DEFAULT_SORT];
-
-	useEffect(() => {
-		if (initialCity !== currentCity) {
-			setCity(initialCity);
-		}
-		if (initialSort !== currenSort) {
-			setSort(initialSort);
-		}
-	}, [initialCity, initialSort, currentCity, currenSort, setCity, setSort]);
 
 	const mainClass = clsx(
 		'page__main page__main--index',
 		{'page__main--index-empty': isEmpty},
 	);
-
 	const containerClass = clsx (
 		'cities__places-container container',
 		{'cities__places-container--empty': isEmpty},
